+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.x.y] - YYYY-MM-DD (Unreleased)
+
+### Added
+
+- Added a `--hyperlinks` flag to render file paths as clickable links in classic `view` mode.
+
 ## [0.2.1] - 2025-06-23
 
 ### Removed